Refactor order management components
- Replaced `MealOrders` with `GetOrders` in `CustomerOrderListPage.vue` and `KitchenPage.vue`.
- Simplified data structure and template in `CustomerOrderListPage.vue`.
- Updated `GetOrders.vue` to include loading state and error handling, displaying orders by status.
- Removed `MealOrders.vue` as its functionality is now covered by `GetOrders`.
- Introduced new components: `GetOrdersCustomer.vue` and `GetOrdersKitchen.vue` for order display.
- Added styles to new components for improved layout and visual distinction.
